Updating the directory upload to use better utilities
