+++ pong/storage.py
@@ -4,6 +4,7 @@ import time
 import subprocess
 import shutil
 from warnings import warn
+from glob import glob
 
 import google.cloud.exceptions
 from google.cloud import storage
@@ -236,13 +237,62 @@ def get_header(img_blob, parse_line=None):
     return headers
 
 
+def upload_directory_to_bucket(pan_id, dir_name, bucket='panoptes-survey', **kwargs):
+    """Upload an observation directory to google cloud storage.
+
+    Note:
+        This requires that the command line utility `gsutil` be installed
+        and that authentication has properly been set up.
+
+    Args:
+        pan_id (str): A string representing the unit id, e.g. PAN001.
+        dir_name (str): Full path to observation directory.
+        bucket (str, optional): The bucket to place the images in, defaults
+            to 'panoptes-survey'.
+        **kwargs: Optional keywords: verbose
+    """
+    assert os.path.exists(dir_name)
+    assert re.match('PAN\d\d\d', pan_id) is not None
+
+    verbose = kwargs.get('verbose', False)
+
+    def _print(msg):
+        if verbose:
+            print(msg)
+
+    dir_name = dir_name.replace('//', '/')
+    _print("Uploading {}".format(dir_name))
+
+    gsutil = shutil.which('gsutil')
+
+    img_path = os.path.join(dir_name, '*.fz')
+    if glob(img_path):
+        field_dir = dir_name.split('/fields/')[-1]
+        remote_path = os.path.normpath(os.path.join(pan_id, field_dir))
+
+        bucket = 'gs://{}/'.format(bucket)
+        # normpath strips the trailing slash so add here so we place in directory
+        run_cmd = [gsutil, '-mq', 'cp', '-r', img_path, bucket + remote_path + '/']
+        _print("Running: {}".format(run_cmd))
+
+        try:
+            completed_process = subprocess.run(run_cmd, stdout=subprocess.PIPE)
+
+            if completed_process.returncode != 0:
+                warn("Problem uploading")
+                warn(completed_process.stdout)
+        except Exception as e:
+            warn("Problem uploading: {}".format(e))
+
+
 def upload_to_bucket(bucket, local_path, remote_path):
     """ Upload to Google Storage Bucket using gsutil.
 
     Note:
         This function does no sanity checking on `remote_path`.
 
-    See also: `upload_fits_file()`
+    See also: `upload_directory` and `upload_fits_file()` for more
+    convenient functions.
 
     Args:
         bucket (str): Bucket name.
